+++ src/Element/Element.php
@@ -23,7 +23,6 @@ use craft\elements\actions\SetStatus;
 use craft\elements\actions\View as ViewAction;
 use craft\elements\conditions\ElementCondition;
 use craft\elements\conditions\ElementConditionInterface;
-use craft\elements\db\EagerLoadPlan;
 use craft\elements\db\NestedElementQueryInterface;
 use craft\elements\exporters\Expanded;
 use craft\elements\exporters\Raw;
@@ -33,14 +32,12 @@ use craft\events\AuthorizationCheckEvent;
 use craft\events\DefineAltActionsEvent;
 use craft\events\DefineAttributeHtmlEvent;
 use craft\events\DefineAttributeKeywordsEvent;
-use craft\events\DefineEagerLoadingMapEvent;
 use craft\events\DefineHtmlEvent;
 use craft\events\DefineMenuItemsEvent;
 use craft\events\DefineMetadataEvent;
 use craft\events\DefineUrlEvent;
 use craft\events\DefineValueEvent;
 use craft\events\ElementIndexTableAttributeEvent;
-use craft\events\ElementStructureEvent;
 use craft\events\ModelEvent;
 use craft\events\RegisterElementActionsEvent;
 use craft\events\RegisterElementCardAttributesEvent;
@@ -55,7 +52,6 @@ use craft\events\RegisterElementSourcesEvent;
 use craft\events\RegisterElementTableAttributesEvent;
 use craft\events\RegisterPreviewTargetsEvent;
 use craft\events\RenderElementEvent;
-use craft\events\SetEagerLoadedElementsEvent;
 use craft\events\SetElementRouteEvent;
 use craft\fieldlayoutelements\BaseField;
 use craft\fieldlayoutelements\CustomField;
@@ -72,6 +68,7 @@ use CraftCms\Cms\Auth\SessionAuth;
 use CraftCms\Cms\Cms;
 use CraftCms\Cms\Database\Table;
 use CraftCms\Cms\Element\Concerns\Draftable;
+use CraftCms\Cms\Element\Concerns\Eagerloadable;
 use CraftCms\Cms\Element\Concerns\Revisionable;
 use CraftCms\Cms\Element\Concerns\Structurable;
 use CraftCms\Cms\Element\Enums\AttributeStatus;
@@ -117,7 +114,6 @@ use ReflectionClass;
 use Stringable;
 use Throwable;
 use Tpetry\QueryExpressions\Function\Conditional\Coalesce;
-use Tpetry\QueryExpressions\Language\Alias;
 use Traversable;
 use Twig\Markup;
 use UnitEnum;
@@ -172,8 +168,6 @@ use function CraftCms\Cms\t;
  * @property int $totalDescendants The total number of descendants that the element has
  * @property string|null $uriFormat The URI format used to generate this element’s URL
  * @property string|null $url The element’s full URL
- *
- * @phpstan-import-type EagerLoadingMap from ElementInterface
  */
 #[Ruleset(ElementRules::class)]
 abstract class Element extends Component implements ElementInterface
@@ -181,7 +175,10 @@ abstract class Element extends Component implements ElementInterface
     use ArrayableTrait {
         toArray as traitToArray;
     }
-    use Draftable;
+    use Draftable {
+        canCreateDrafts as traitCanCreateDrafts;
+    }
+    use Eagerloadable;
     use ElementTrait;
     use Macroable {
         __call as macroCall;
@@ -352,50 +349,6 @@ abstract class Element extends Component implements ElementInterface
      * @since 5.5.0
      */
     public const EVENT_REGISTER_DEFAULT_CARD_ATTRIBUTES = 'registerDefaultCardAttributes';
-
-    /**
-     * @event DefineEagerLoadingMapEvent The event that is triggered when defining an eager-loading map.
-     *
-     * ```php
-     * use CraftCms\Cms\Element\Element;
-     * use craft\base\ElementInterface;
-     * use craft\db\Query;
-     * use CraftCms\Cms\Entry\Elements\Entry;
-     * use craft\events\DefineEagerLoadingMapEvent;
-     * use yii\base\Event;
-     *
-     * // Add support for `with(['bookClub'])` to entries
-     * Event::on(
-     *     Entry::class,
-     *     Element::EVENT_DEFINE_EAGER_LOADING_MAP,
-     *     function(DefineEagerLoadingMapEvent $event) {
-     *         if ($event->handle === 'bookClub') {
-     *             $bookEntryIds = array_map(fn(ElementInterface $element) => $element->id, $event->elements);
-     *             $event->elementType = \my\plugin\BookClub::class,
-     *             $event->map = (new Query)
-     *                 ->select(['source' => 'bookId', 'target' => 'clubId'])
-     *                 ->from('{{%bookclub_books}}')
-     *                 ->where(['bookId' => $bookEntryIds])
-     *                 ->all();
-     *             $event->handled = true;
-     *         }
-     *     }
-     * );
-     * ```
-     *
-     * @since 3.1.0
-     */
-    public const EVENT_DEFINE_EAGER_LOADING_MAP = 'defineEagerLoadingMap';
-
-    /**
-     * @event SetEagerLoadedElementsEvent The event that is triggered when setting eager-loaded elements.
-     *
-     * Set [[Event::$handled]] to `true` to prevent the elements from getting stored to the private
-     * `$_eagerLoadedElements` array.
-     *
-     * @since 3.5.0
-     */
-    public const EVENT_SET_EAGER_LOADED_ELEMENTS = 'setEagerLoadedElements';
 
     /**
      * @event RegisterPreviewTargetsEvent The event that is triggered when registering the element’s preview targets.
@@ -1815,453 +1768,6 @@ abstract class Element extends Component implements ElementInterface
 
     /**
      * {@inheritdoc}
-     *
-     * @return EagerLoadingMap|null|false
-     */
-    public static function eagerLoadingMap(array $sourceElements, string $handle): array|null|false
-    {
-        switch ($handle) {
-            case 'descendants':
-            case 'children':
-                return self::_mapDescendants($sourceElements, $handle === 'children');
-            case 'ancestors':
-            case 'parent':
-                return self::_mapAncestors($sourceElements, $handle === 'parent');
-            case 'localized':
-                return self::_mapLocalized($sourceElements);
-            case 'currentRevision':
-                return self::_mapCurrentRevisions($sourceElements);
-            case 'drafts':
-                return self::_mapDrafts($sourceElements);
-            case 'revisions':
-                return self::_mapRevisions($sourceElements);
-            case 'draftCreator':
-                return self::_mapDraftCreators($sourceElements);
-            case 'revisionCreator':
-                return self::_mapRevisionCreators($sourceElements);
-        }
-
-        // Is $handle a custom field handle?
-        // (Leave it up to the extended class to set the field context, if it shouldn't be 'global')
-        if (str_contains($handle, ':')) {
-            [$providerHandle, $fieldHandle] = explode(':', $handle, 2);
-        } else {
-            $providerHandle = null;
-            $fieldHandle = $handle;
-        }
-
-        // Get all the custom fields by that handle
-        $fields = [];
-        foreach (static::fieldLayouts(null) as $fieldLayout) {
-            if ($providerHandle === null || $fieldLayout->provider?->getHandle() === $providerHandle) {
-                $layoutField = $fieldLayout->getFieldByHandle($fieldHandle);
-                if ($layoutField) {
-                    $fields[] = $layoutField;
-                    if ($providerHandle !== null) {
-                        break;
-                    }
-                }
-            }
-        }
-
-        // If there were any matching fields, find the first one that's actually included in
-        // at least one of the source elements’ field layouts
-        if (! empty($fields)) {
-            foreach ($fields as $field) {
-                if (! $field instanceof EagerLoadingFieldInterface) {
-                    continue;
-                }
-
-                // filter out elements, if field is not part of its layout
-                // https://github.com/craftcms/cms/issues/12539
-                $fieldSourceElements = array_values(
-                    array_filter($sourceElements, function ($sourceElement) use ($field) {
-                        $layoutField = $sourceElement->getFieldLayout()?->getFieldByHandle($field->handle);
-
-                        return $layoutField && $layoutField->id === $field->id;
-                    }),
-                );
-
-                if (! empty($fieldSourceElements)) {
-                    return $field->getEagerLoadingMap($fieldSourceElements);
-                }
-            }
-
-            // None of the source elements include any of the matching fields
-            return false;
-        }
-
-        // Fire a 'defineEagerLoadingMap' event
-        if (Event::hasHandlers(static::class, self::EVENT_DEFINE_EAGER_LOADING_MAP)) {
-            $event = new DefineEagerLoadingMapEvent([
-                'sourceElements' => $sourceElements,
-                'handle' => $handle,
-            ]);
-            Event::trigger(static::class, self::EVENT_DEFINE_EAGER_LOADING_MAP, $event);
-            if ($event->elementType !== null) {
-                return [
-                    'elementType' => $event->elementType,
-                    'map' => $event->map,
-                    'criteria' => $event->criteria,
-                ];
-            }
-        }
-
-        // return null so eager-loading is ignored for this handle
-        return null;
-    }
-
-    /**
-     * Returns an eager-loading map for the source elements’ descendants.
-     *
-     * @param  ElementInterface[]  $sourceElements  An array of the source elements
-     * @param  bool  $children  Whether only direct children should be included
-     * @return array|null The eager-loading element ID mappings, or null if the result should be ignored
-     */
-    private static function _mapDescendants(array $sourceElements, bool $children): ?array
-    {
-        $elementStructureData = self::_structureDataForElements($sourceElements, $children);
-
-        if (empty($elementStructureData)) {
-            return null;
-        }
-
-        // Build the descendant condition & params
-        $descendantStructureQuery = DB::table(Table::STRUCTUREELEMENTS);
-
-        foreach ($elementStructureData as $elementStructureDatum) {
-            $descendantStructureQuery->orWhere(function (Builder $query) use ($children, $elementStructureDatum) {
-                $query->where('structureId', $elementStructureDatum['structureId'])
-                    ->where('lft', '>', $elementStructureDatum['lft'])
-                    ->where('rgt', '<', $elementStructureDatum['rgt'])
-                    ->when($children, fn (Builder $query) => $query->where('level', $elementStructureDatum['level'] + 1));
-            });
-        }
-
-        // Fetch the descendant data
-        $descendantStructureQuery
-            ->select(['structureId', 'lft', 'rgt', 'elementId'])
-            ->when($children, fn (Builder $query) => $query->addSelect('level'))
-            ->orderBy('lft');
-
-        $descendantStructureData = $descendantStructureQuery
-            ->get()
-            ->map(fn (object $data) => (array) $data);
-
-        // Map the elements to their descendants
-        $map = [];
-        foreach ($elementStructureData as $elementStructureDatum) {
-            foreach ($descendantStructureData as $descendantStructureDatum) {
-                if (! ($descendantStructureDatum['structureId'] == $elementStructureDatum['structureId'] && $descendantStructureDatum['lft'] > $elementStructureDatum['lft'] && $descendantStructureDatum['rgt'] < $elementStructureDatum['rgt'])) {
-                    continue;
-                }
-                if (! (! $children || $descendantStructureDatum['level'] == $elementStructureDatum['level'] + 1)) {
-                    continue;
-                }
-                if (! $descendantStructureDatum['elementId']) {
-                    continue;
-                }
-                $map[] = [
-                    'source' => $elementStructureDatum['elementId'],
-                    'target' => $descendantStructureDatum['elementId'],
-                ];
-            }
-        }
-
-        return [
-            'elementType' => static::class,
-            'map' => $map,
-        ];
-    }
-
-    /**
-     * Returns an eager-loading map for the source elements’ ancestors.
-     *
-     * @param  ElementInterface[]  $sourceElements  An array of the source elements
-     * @param  bool  $parents  Whether only direct parents should be included
-     * @return array|null The eager-loading element ID mappings, or null if the result should be ignored
-     */
-    private static function _mapAncestors(array $sourceElements, bool $parents): ?array
-    {
-        $elementStructureData = self::_structureDataForElements($sourceElements, $parents);
-
-        if (empty($elementStructureData)) {
-            return null;
-        }
-
-        $ancestorStructureQuery = DB::table(Table::STRUCTUREELEMENTS);
-
-        // Build the ancestor condition & params
-        foreach ($elementStructureData as $elementStructureDatum) {
-            $ancestorStructureQuery->orWhere(function (Builder $query) use ($elementStructureDatum, $parents) {
-                $query->where('structureId', $elementStructureDatum['structureId'])
-                    ->where('lft', '<', $elementStructureDatum['lft'])
-                    ->where('rgt', '>', $elementStructureDatum['rgt'])
-                    ->when($parents, fn (Builder $query) => $query->where('level', $elementStructureDatum['level'] - 1));
-            });
-        }
-
-        // Fetch the ancestor data
-        $ancestorStructureQuery
-            ->select(['structureId', 'lft', 'rgt', 'elementId'])
-            ->when($parents, fn (Builder $query) => $query->addSelect('level'))
-            ->orderBy('lft');
-
-        $ancestorStructureData = $ancestorStructureQuery
-            ->get()
-            ->map(fn (object $data) => (array) $data);
-
-        // Map the elements to their ancestors
-        $map = [];
-        foreach ($elementStructureData as $elementStructureDatum) {
-            foreach ($ancestorStructureData as $ancestorStructureDatum) {
-                if (
-                    $ancestorStructureDatum['structureId'] == $elementStructureDatum['structureId'] &&
-                    $ancestorStructureDatum['lft'] < $elementStructureDatum['lft'] &&
-                    $ancestorStructureDatum['rgt'] > $elementStructureDatum['rgt'] &&
-                    (! $parents || $ancestorStructureDatum['level'] == $elementStructureDatum['level'] - 1)
-                ) {
-                    if ($ancestorStructureDatum['elementId']) {
-                        $map[] = [
-                            'source' => $elementStructureDatum['elementId'],
-                            'target' => $ancestorStructureDatum['elementId'],
-                        ];
-                    }
-
-                    // If we're just fetching the parents, then we're done with this element
-                    if ($parents) {
-                        break;
-                    }
-                }
-            }
-        }
-
-        return [
-            'elementType' => static::class,
-            'map' => $map,
-        ];
-    }
-
-    /**
-     * @param  ElementInterface[]  $elements
-     */
-    private static function _structureDataForElements(array $elements, bool $withLevel): array
-    {
-        $data = [];
-        $fetchDataForIds = [];
-
-        foreach ($elements as $element) {
-            if (isset($element->structureId, $element->lft, $element->rgt, $element->level)) {
-                $data[] = [
-                    'structureId' => $element->structureId,
-                    'elementId' => $element->id,
-                    'lft' => $element->lft,
-                    'rgt' => $element->rgt,
-                    'level' => $element->level,
-                ];
-            } else {
-                $fetchDataForIds[] = $element->id;
-            }
-        }
-
-        if (! empty($fetchDataForIds)) {
-            $fetched = DB::table(Table::STRUCTUREELEMENTS)
-                ->whereIn('elementId', $fetchDataForIds)
-                ->select(['structureId', 'elementId', 'lft', 'rgt'])
-                ->when($withLevel, fn (Builder $query) => $query->addSelect('level'))
-                ->get()
-                ->map(fn (object $data) => (array) $data)
-                ->all();
-
-            array_push($data, ...$fetched);
-        }
-
-        return $data;
-    }
-
-    /**
-     * Returns an eager-loading map for the source elements in other locales.
-     *
-     * @param  ElementInterface[]  $sourceElements  An array of the source elements
-     * @return array The eager-loading element ID mappings
-     */
-    private static function _mapLocalized(array $sourceElements): array
-    {
-        $sourceSiteId = $sourceElements[0]->siteId;
-        $otherSiteIds = [];
-        foreach (Sites::getAllSites() as $site) {
-            if ($site->id != $sourceSiteId) {
-                $otherSiteIds[] = $site->id;
-            }
-        }
-
-        // Map the source elements to themselves
-        $map = [];
-        if (! empty($otherSiteIds)) {
-            foreach ($sourceElements as $element) {
-                $map[] = [
-                    'source' => $element->id,
-                    'target' => $element->id,
-                ];
-            }
-        }
-
-        return [
-            'elementType' => static::class,
-            'map' => $map,
-            'criteria' => [
-                'siteId' => $otherSiteIds,
-                'drafts' => null,
-                'provisionalDrafts' => null,
-                'revisions' => null,
-            ],
-        ];
-    }
-
-    /**
-     * Returns an eager-loading map for the source elements’ current revisions.
-     *
-     * @param  ElementInterface[]  $sourceElements  An array of the source elements
-     * @return array The eager-loading element ID mappings
-     */
-    private static function _mapCurrentRevisions(array $sourceElements): array
-    {
-        // Get the source element IDs
-        $sourceElementIds = array_map(fn (ElementInterface $element) => $element->id, $sourceElements);
-
-        $map = DB::table(Table::ELEMENTS, 're')
-            ->join(new Alias(Table::REVISIONS, 'r'), 'r.id', '=', 're.revisionId')
-            ->join(new Alias(Table::ELEMENTS, 'se'), 'se.id', '=', 'r.canonicalId')
-            ->whereColumn('re.dateCreated', '=', 'se.dateUpdated')
-            ->whereIn('se.id', $sourceElementIds)
-            ->select([
-                'se.id as source',
-                're.id as target',
-            ])
-            ->get()
-            ->map(fn (object $data) => (array) $data);
-
-        return [
-            'elementType' => static::class,
-            'map' => $map,
-            'criteria' => ['revisions' => true, 'status' => null],
-        ];
-    }
-
-    /**
-     * Returns an eager-loading map for the source elements’ current drafts.
-     *
-     * @param  ElementInterface[]  $sourceElements  An array of the source elements
-     * @return array The eager-loading element ID mappings
-     */
-    private static function _mapDrafts(array $sourceElements): array
-    {
-        // Get the source element IDs
-        $sourceElementIds = array_map(fn (ElementInterface $element) => $element->id, $sourceElements);
-
-        $map = DB::table(Table::DRAFTS, 'd')
-            ->join(new Alias(Table::ELEMENTS, 'e'), 'e.draftId', '=', 'd.id')
-            ->whereIn('d.canonicalId', $sourceElementIds)
-            ->select([
-                'd.canonicalId as source',
-                'e.id as target',
-            ])
-            ->get()
-            ->map(fn (object $data) => (array) $data);
-
-        return [
-            'elementType' => static::class,
-            'map' => $map,
-            'criteria' => ['drafts' => true],
-        ];
-    }
-
-    /**
-     * Returns an eager-loading map for the source elements’ current revisions.
-     *
-     * @param  ElementInterface[]  $sourceElements  An array of the source elements
-     * @return array The eager-loading element ID mappings
-     */
-    private static function _mapRevisions(array $sourceElements): array
-    {
-        // Get the source element IDs
-        $sourceElementIds = array_map(fn (ElementInterface $element) => $element->id, $sourceElements);
-
-        $map = DB::table(Table::REVISIONS, 'r')
-            ->join(new Alias(Table::ELEMENTS, 'e'), 'e.revisionId', '=', 'r.id')
-            ->whereIn('r.canonicalId', $sourceElementIds)
-            ->select([
-                'r.canonicalId as source',
-                'e.id as target',
-            ])
-            ->get()
-            ->map(fn (object $data) => (array) $data);
-
-        return [
-            'elementType' => static::class,
-            'map' => $map,
-            'criteria' => ['revisions' => true],
-        ];
-    }
-
-    /**
-     * Returns an eager-loading map for the source elements’ draft creators.
-     *
-     * @param  ElementInterface[]  $sourceElements  An array of the source elements
-     * @return array The eager-loading element ID mappings
-     */
-    private static function _mapDraftCreators(array $sourceElements): array
-    {
-        // Get the source element IDs
-        $sourceElementIds = array_map(fn (ElementInterface $element) => $element->id, $sourceElements);
-
-        $map = DB::table(Table::ELEMENTS, 'e')
-            ->join(new Alias(Table::DRAFTS, 'd'), 'd.id', '=', 'e.draftId')
-            ->whereIn('e.id', $sourceElementIds)
-            ->whereNotNull('d.creatorId')
-            ->select([
-                'e.id as source',
-                'd.creatorId as target',
-            ])
-            ->get()
-            ->map(fn (object $data) => (array) $data);
-
-        return [
-            'elementType' => User::class,
-            'map' => $map,
-        ];
-    }
-
-    /**
-     * Returns an eager-loading map for the source elements’ revision creators.
-     *
-     * @param  ElementInterface[]  $sourceElements  An array of the source elements
-     * @return array The eager-loading element ID mappings
-     */
-    private static function _mapRevisionCreators(array $sourceElements): array
-    {
-        // Get the source element IDs
-        $sourceElementIds = array_map(fn (ElementInterface $element) => $element->id, $sourceElements);
-
-        $map = DB::table(Table::ELEMENTS, 'e')
-            ->join(new Alias(Table::REVISIONS, 'r'), 'r.id', '=', 'e.revisionId')
-            ->whereIn('e.id', $sourceElementIds)
-            ->whereNotNull('r.creatorId')
-            ->select([
-                'e.id as source',
-                'r.creatorId as target',
-            ])
-            ->get()
-            ->map(fn (object $data) => (array) $data);
-
-        return [
-            'elementType' => User::class,
-            'map' => $map,
-        ];
-    }
-
-    /**
-     * {@inheritdoc}
      */
     public static function baseGqlType(): Type
     {
@@ -2469,30 +1975,6 @@ abstract class Element extends Component implements ElementInterface
      * @see addInvalidNestedElementIds()
      */
     private array $_invalidNestedElementIds = [];
-
-    /**
-     * @var array<string,ElementCollection>
-     *
-     * @see getEagerLoadedElements()
-     * @see setEagerLoadedElements()
-     */
-    private array $_eagerLoadedElements = [];
-
-    /**
-     * @var array<string,bool>
-     *
-     * @see getFieldValue()
-     * @see setLazyEagerLoadedElements()
-     */
-    private array $_lazyEagerLoadedElements = [];
-
-    /**
-     * @var array<string,int>
-     *
-     * @see getEagerLoadedElementCount()
-     * @see setEagerLoadedElementCount
-     */
-    private array $_eagerLoadedElementCounts = [];
 
     /**
      * @var bool|bool[]
@@ -4906,123 +4388,6 @@ JS, [
     public function addInvalidNestedElementIds(array $ids): void
     {
         array_push($this->_invalidNestedElementIds, ...$ids);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function hasEagerLoadedElements(string $handle): bool
-    {
-        if (! isset($this->_eagerLoadedElements[$handle])) {
-            // See if we have it stored with the field layout provider’s handle
-            $providerHandle = $this->providerHandle();
-            if ($providerHandle !== null && isset($this->_eagerLoadedElements["$providerHandle:$handle"])) {
-                $handle = "$providerHandle:$handle";
-            }
-        }
-
-        return isset($this->_eagerLoadedElements[$handle]);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getEagerLoadedElements(string $handle): ?ElementCollection
-    {
-        if (! isset($this->_eagerLoadedElements[$handle])) {
-            // See if we have it stored with the field layout provider’s handle
-            $providerHandle = $this->providerHandle();
-            if ($providerHandle !== null && isset($this->_eagerLoadedElements["$providerHandle:$handle"])) {
-                $handle = "$providerHandle:$handle";
-            } else {
-                return null;
-            }
-        }
-
-        $elements = $this->_eagerLoadedElements[$handle];
-        ElementHelper::setNextPrevOnElements($elements);
-
-        return $elements;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setEagerLoadedElements(string $handle, array $elements, EagerLoadPlan $plan): void
-    {
-        switch ($plan->handle) {
-            case 'parent':
-                $this->_parent = $elements[0] ?? false;
-                break;
-            case 'currentRevision':
-                $this->currentRevision = $elements[0] ?? false;
-                break;
-            case 'draftCreator':
-                /** @var User[] $elements */
-                $this->setDraftCreator($elements[0] ?? null);
-                break;
-            case 'revisionCreator':
-                /** @var User[] $elements */
-                $this->setRevisionCreator($elements[0] ?? null);
-                break;
-            default:
-                // Fire a 'setEagerLoadedElements' event
-                if ($this->hasEventHandlers(self::EVENT_SET_EAGER_LOADED_ELEMENTS)) {
-                    $event = new SetEagerLoadedElementsEvent([
-                        'handle' => $handle,
-                        'elements' => $elements,
-                        'plan' => $plan,
-                    ]);
-                    $this->trigger(self::EVENT_SET_EAGER_LOADED_ELEMENTS, $event);
-                    if ($event->handled) {
-                        break;
-                    }
-                }
-
-                // No takers. Just store it in the internal array then.
-                $this->_eagerLoadedElements[$handle] = ElementCollection::make($elements);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLazyEagerLoadedElements(string $handle, bool $value = true): void
-    {
-        $this->_lazyEagerLoadedElements[$handle] = $value;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getEagerLoadedElementCount(string $handle): ?int
-    {
-        if (! isset($this->_eagerLoadedElementCounts[$handle])) {
-            // See if we have it stored with the field layout provider’s handle
-            $providerHandle = $this->providerHandle();
-            if ($providerHandle !== null && isset($this->_eagerLoadedElements["$providerHandle:$handle"])) {
-                $handle = "$providerHandle:$handle";
-            }
-        }
-
-        return $this->_eagerLoadedElementCounts[$handle] ?? null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setEagerLoadedElementCount(string $handle, int $count): void
-    {
-        $this->_eagerLoadedElementCounts[$handle] = $count;
-    }
-
-    private function providerHandle(): ?string
-    {
-        try {
-            return $this->getFieldLayout()?->provider?->getHandle();
-        } catch (InvalidConfigException) {
-            return null;
-        }
     }
 
     /**
